feat(endpoints): share management kebab with CF picker, add Reconnect
The /cloud-foundry picker is a projection of the Endpoints page onto its
CF subset, but its cards carried no management actions. The per-endpoint
kebab (Connect / Disconnect / Edit / Unregister) moves out of the
endpoints signal list into a shared EndpointRowActionsService consumed
by both surfaces, so the menus cannot drift. The picker passes
unregister:false - registering and unregistering are admin-level
operations on the global registration and stay on the Endpoints page.

Adds the long-discussed Reconnect action for connected endpoints:
re-auth in place via the connect dialog, no manual disconnect step -
the same flow the stale-token snackbar uses. Two dialog fixes make
that flow safe:

- the connect success detector derived 'connected' from endpoint.user
  being set, which is true from the moment the dialog opens against an
  already-connected endpoint - after the 1s connect delay the dialog
  self-closed with a false 'Connected' snackbar. Success is now gated
  on hasAttemptedConnect, the same guard the error path already used.
- the username prefill only knew the localStorage history of past
  connects; the dialog config now carries the live connection's user,
  preferred over the remembered value, so Reconnect opens prefilled.

Live-verified: reconnect round-trip against a dev instance replaces
the token with the dialog staying open until submit.
