+++ src/frontend/packages/core/src/features/endpoints/connect-endpoint/connect-endpoint.component.ts
@@ -124,11 +124,12 @@ export class ConnectEndpointComponent implements OnInit, OnDestroy {
     });
     // Add authValues as a separate group to handle dynamic auth type switching
     this.endpointForm.addControl('authValues', this.fb.group(this.autoSelected.form || {}));
-    // Prefill the username from the last successful connect (stored in
-    // localStorage by connect.service on success) — backend EndpointModel
-    // clears endpoint.user on disconnect, so the dialog otherwise opens
-    // empty even when the same user is reconnecting.
-    this.prefillRememberedUsername(config.guid);
+    // Prefill the username: prefer the live connection's user (supplied for
+    // the Reconnect flow), falling back to the last successful connect
+    // remembered in localStorage — backend EndpointModel clears
+    // endpoint.user on disconnect, so the dialog otherwise opens empty even
+    // when the same user is reconnecting.
+    this.prefillUsername(config);
     this.authChanged();
 
     // Template container reference is not available at construction
@@ -204,14 +205,14 @@ export class ConnectEndpointComponent implements OnInit, OnDestroy {
     return a.length === b.length && a.filter(item => b.indexOf(item) < 0).length === 0;
   }
 
-  private prefillRememberedUsername(endpointGuid: string): void {
+  private prefillUsername(config: ConnectEndpointConfig): void {
     const authValues = this.endpointForm.get('authValues');
     if (!authValues || !authValues.get('username')) {
       return;
     }
-    const stored = getRememberedUsername(endpointGuid);
-    if (stored) {
-      authValues.patchValue({ username: stored });
+    const username = config.username || getRememberedUsername(config.guid);
+    if (username) {
+      authValues.patchValue({ username });
     }
   }
