+++ src/frontend/packages/core/src/public-api.ts
@@ -252,6 +252,7 @@ export { CreateEndpointModule } from './features/endpoints/create-endpoint/creat
 export { CreateEndpointBaseStepComponent } from './features/endpoints/create-endpoint/create-endpoint-base-step/create-endpoint-base-step.component';
 export { ConnectEndpointDialogComponent } from './features/endpoints/connect-endpoint-dialog/connect-endpoint-dialog.component';
 export { ConnectEndpointConfig } from './features/endpoints/connect.service';
+export { EndpointRowActionsService } from './features/endpoints/endpoint-row-actions.service';
 export { BaseEndpointAuth, EndpointAuthTypeNames } from './core/endpoint-auth';
 
 // List Types and Components
